tratamento de erros e testes unitarios

+++ payment-service/src/main/java/br/com/fiap/paymentservice/restcontroller/PaymentController.java
@@ -1,6 +1,9 @@
 package br.com.fiap.paymentservice.restcontroller;
 
+import br.com.fiap.paymentservice.exceptions.BadRequestException;
+import br.com.fiap.paymentservice.exceptions.NotFoundException;
 import br.com.fiap.paymentservice.model.Payment;
+import br.com.fiap.paymentservice.model.ResponseException;
 import br.com.fiap.paymentservice.repository.PaymentRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -10,6 +13,7 @@ import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 
 import java.math.BigInteger;
 import java.net.URI;
+import java.time.LocalDateTime;
 
 @RestController
 @RequestMapping("/payment")
@@ -21,15 +25,26 @@ public class PaymentController {
     @GetMapping("/findById/{id}")
     public ResponseEntity findById(@PathVariable String id) {
 
+        int compare = id.compareTo( String.valueOf( new BigInteger("10") ) );
+
+        if (compare == 1) {
+            return new BadRequestException().handleAllExceptions(new ResponseException( LocalDateTime.now(), "Parametro id náo encontrado", HttpStatus.BAD_REQUEST));
+        }
+
         Payment payment = paymentRepository.findById(id);
         if (payment != null) {
             return new ResponseEntity(payment, HttpStatus.OK);
+        } else {
+            throw new NotFoundException( "payment with id ".concat( id ).concat( " not found" ) );
         }
-        return new ResponseEntity(null, HttpStatus.OK);
     }
 
     @PostMapping("/save")
     public ResponseEntity save(@RequestBody Payment payment) {
+
+        if (payment.getTransactionId() == null) {
+            throw new NotFoundException("id of Order not passed");
+        }
         paymentRepository.save(payment);
         URI location = ServletUriComponentsBuilder.fromCurrentRequest().path("/id").buildAndExpand(payment.getTransactionId()).toUri();
         return ResponseEntity.created(location).build();
@@ -37,12 +52,21 @@ public class PaymentController {
 
     @PutMapping("/update/{id}")
     public ResponseEntity update(@RequestBody Payment payment, @PathVariable String id) {
+
+        Payment paymentDatabase = paymentRepository.findById( id );
+        if (paymentDatabase == null) {
+            throw new NotFoundException("Payment not found");
+        }
         Payment response = paymentRepository.update(payment, id);
         return new ResponseEntity(response, HttpStatus.OK);
     }
 
     @DeleteMapping("/delete/{id}")
     public ResponseEntity delete(@PathVariable  String id) {
+        Payment paymentDatabase = paymentRepository.findById( id );
+        if (paymentDatabase == null) {
+            throw new NotFoundException("Payment not found");
+        }
         paymentRepository.delete(id);
         return new ResponseEntity(HttpStatus.OK);
     }
